Moved files around in the graphql packge
