+++ packages/graphql/index.ts
@@ -1,13 +1,4 @@
-export * from './generated/types'
-
-export * from './measurePerformanceLink'
-export * from './errorLink'
-
-export { default as fragments } from './generated/fragments.json'
-
-export * from './migrateCache'
-export * from './typePolicies'
-export * from './createCacheReviver'
-
 export * from './apollo'
 export * from './components/GraphQLProvider'
+export { default as fragments } from './generated/fragments.json'
+export * from './generated/types'
